+++ src/app/Vitessce.js
@@ -6,7 +6,10 @@ import {
 } from '@material-ui/core/styles';
 import isEqual from 'lodash/isEqual';
 import { muiTheme } from '../components/shared-mui/styles';
-import { ViewConfigProvider, createViewConfigStore } from './state/hooks';
+import {
+  ViewConfigProvider, createViewConfigStore,
+  AuxiliaryProvider, createAuxiliaryStore,
+} from './state/hooks';
 
 import VitessceGrid from './VitessceGrid';
 import Warning from './Warning';
@@ -91,19 +94,21 @@ export default function Vitessce(props) {
     <StylesProvider generateClassName={generateClassName}>
       <ThemeProvider theme={muiTheme[theme]}>
         <ViewConfigProvider createStore={createViewConfigStore}>
-          <VitessceGrid
-            config={configOrWarning}
-            getComponent={getComponent}
-            rowHeight={rowHeight}
-            height={height}
-            theme={theme}
-          />
-          <CallbackPublisher
-            onWarn={onWarn}
-            onConfigChange={onConfigChange}
-            onLoaderChange={onLoaderChange}
-            validateOnConfigChange={validateOnConfigChange}
-          />
+          <AuxiliaryProvider createStore={createAuxiliaryStore}>
+            <VitessceGrid
+              config={configOrWarning}
+              getComponent={getComponent}
+              rowHeight={rowHeight}
+              height={height}
+              theme={theme}
+            />
+            <CallbackPublisher
+              onWarn={onWarn}
+              onConfigChange={onConfigChange}
+              onLoaderChange={onLoaderChange}
+              validateOnConfigChange={validateOnConfigChange}
+            />
+          </AuxiliaryProvider>
         </ViewConfigProvider>
       </ThemeProvider>
     </StylesProvider>