+++ src/app/VitessceGrid.js
@@ -4,7 +4,7 @@ import React, {
 import { VitessceGridLayout } from './vitessce-grid-layout';
 import { useRowHeight, createLoaders } from './vitessce-grid-utils';
 import {
-  useViewConfigStore,
+  useViewConfigStoreApi,
   useSetViewConfig,
   useSetLoaders,
   useEmitGridResize,
@@ -44,7 +44,8 @@ export default function VitessceGrid(props) {
     onResize();
   }, [rowHeight, onResize]);
 
-  const setViewConfig = useSetViewConfig(useViewConfigStore);
+  const viewConfigStoreApi = useViewConfigStoreApi();
+  const setViewConfig = useSetViewConfig(viewConfigStoreApi);
   const setLoaders = useSetLoaders();
   const removeComponent = useRemoveComponent();
   const changeLayout = useChangeLayout();