+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Don't show image layer buttons if there is only one layer.
 - Fix spatial options to only show what is necessary and display at all if necessary.
 - Fix bug introduced by #1037 that broke channel removal/addition.
+- Fix setting default schema values for properties that are not in the current deck.gl view state (for example, the z direction for `target`).
 
 ## [1.1.14](https://www.npmjs.com/package/vitessce/v/1.1.14) - 2021-09-01