+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 - Fix bitmask picking/highlighting.
-
+- Upgrade `Viv` to 0.11.0
 ## [1.1.15](https://www.npmjs.com/package/vitessce/v/1.1.15) - 2021-09-21
 
 ### Added